move rewriteConfig and applyRewriter to ConfigUtil
