+++ samza-core/src/test/java/org/apache/samza/util/TestUtil.java
@@ -6,14 +6,9 @@ import java.net.NetworkInterface;
 import java.net.UnknownHostException;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 import com.google.common.collect.ImmutableMap;
-import org.apache.samza.SamzaException;
 import org.apache.samza.config.ApplicationConfig;
 import org.apache.samza.config.Config;
-import org.apache.samza.config.ConfigRewriter;
-import org.apache.samza.config.JobConfig;
 import org.apache.samza.config.MapConfig;
 import org.apache.samza.config.TaskConfig;
 import org.junit.Test;
@@ -31,12 +26,6 @@ import static org.powermock.api.mockito.PowerMockito.when;
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({Util.NetworkingUtil.class})
 public class TestUtil {
-  private static final String CONFIG_KEY = "config.key";
-  private static final String CONFIG_VALUE = "value";
-  private static final String NEW_CONFIG_KEY = "new.rewritten.config.key";
-  private static final String REWRITER_NAME = "propertyRewriter";
-  private static final String OTHER_REWRITER_NAME = "otherPropertyRewriter";
-
   @Test
   public void testEnvVarEscape() {
     // no special characters in original
@@ -180,110 +169,6 @@ public class TestUtil {
     assertEquals(finalInetAddress, Util.getLocalHost());
   }
 
-  @Test
-  public void testRewriteConfig() {
-    Map<String, String> baseConfigMap = ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE);
-
-    // no rewriters
-    Map<String, String> fullConfig = new HashMap<>(baseConfigMap);
-    assertEquals(fullConfig, Util.rewriteConfig(new MapConfig(fullConfig)));
-
-    // one rewriter
-    fullConfig = new HashMap<>(baseConfigMap);
-    fullConfig.put(JobConfig.CONFIG_REWRITERS, REWRITER_NAME);
-    fullConfig.put(String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME), NewPropertyRewriter.class.getName());
-    Map<String, String> expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.put(NEW_CONFIG_KEY, CONFIG_VALUE);
-    assertEquals(new MapConfig(expectedConfigMap), Util.rewriteConfig(new MapConfig(fullConfig)));
-
-    // only apply rewriters from rewriters list
-    fullConfig = new HashMap<>(baseConfigMap);
-    fullConfig.put(JobConfig.CONFIG_REWRITERS, OTHER_REWRITER_NAME);
-    fullConfig.put(String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME), NewPropertyRewriter.class.getName());
-    fullConfig.put(String.format(JobConfig.CONFIG_REWRITER_CLASS, OTHER_REWRITER_NAME),
-        UpdatePropertyRewriter.class.getName());
-    expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.put(CONFIG_KEY, CONFIG_VALUE + CONFIG_VALUE);
-    assertEquals(new MapConfig(expectedConfigMap), Util.rewriteConfig(new MapConfig(fullConfig)));
-
-    // two rewriters; second rewriter overwrites configs from first
-    fullConfig = new HashMap<>(baseConfigMap);
-    fullConfig.put(JobConfig.CONFIG_REWRITERS, REWRITER_NAME + "," + OTHER_REWRITER_NAME);
-    fullConfig.put(String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME), NewPropertyRewriter.class.getName());
-    fullConfig.put(String.format(JobConfig.CONFIG_REWRITER_CLASS, OTHER_REWRITER_NAME),
-        UpdatePropertyRewriter.class.getName());
-    expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.put(NEW_CONFIG_KEY, CONFIG_VALUE + CONFIG_VALUE);
-    assertEquals(new MapConfig(expectedConfigMap), Util.rewriteConfig(new MapConfig(fullConfig)));
-  }
-
-  /**
-   * This fails because Util will interpret the empty string value as a single rewriter which has the empty string as a
-   * name, and there is no rewriter class config for a rewriter name which is the empty string.
-   * TODO: should this be fixed to interpret the empty string as an empty list?
-   */
-  @Test(expected = SamzaException.class)
-  public void testRewriteConfigConfigRewritersEmptyString() {
-    Config config = new MapConfig(ImmutableMap.of(JobConfig.CONFIG_REWRITERS, ""));
-    Util.rewriteConfig(config);
-  }
-
-  @Test(expected = SamzaException.class)
-  public void testRewriteConfigNoClassForConfigRewriterName() {
-    Config config =
-        new MapConfig(ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, JobConfig.CONFIG_REWRITERS, "unknownRewriter"));
-    Util.rewriteConfig(config);
-  }
-
-  @Test(expected = SamzaException.class)
-  public void testRewriteConfigRewriterClassDoesNotExist() {
-    Config config = new MapConfig(ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, JobConfig.CONFIG_REWRITERS, REWRITER_NAME,
-        String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME), "not_a_class"));
-    Util.rewriteConfig(config);
-  }
-
-  @Test
-  public void testApplyRewriter() {
-    // new property
-    Map<String, String> fullConfig =
-        ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME),
-            NewPropertyRewriter.class.getName());
-    Map<String, String> expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.put(NEW_CONFIG_KEY, CONFIG_VALUE);
-    assertEquals(new MapConfig(expectedConfigMap), Util.applyRewriter(new MapConfig(fullConfig), REWRITER_NAME));
-
-    // update property
-    fullConfig =
-        ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME),
-            UpdatePropertyRewriter.class.getName());
-    expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.put(CONFIG_KEY, CONFIG_VALUE + CONFIG_VALUE);
-    assertEquals(new MapConfig(expectedConfigMap), Util.applyRewriter(new MapConfig(fullConfig), REWRITER_NAME));
-
-    // remove property
-    fullConfig =
-        ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME),
-            DeletePropertyRewriter.class.getName());
-    expectedConfigMap = new HashMap<>(fullConfig);
-    expectedConfigMap.remove(CONFIG_KEY);
-    assertEquals(new MapConfig(expectedConfigMap), Util.applyRewriter(new MapConfig(fullConfig), REWRITER_NAME));
-  }
-
-  @Test(expected = SamzaException.class)
-  public void testApplyRewriterNoClassForConfigRewriterName() {
-    Map<String, String> fullConfig = ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE);
-    Util.applyRewriter(new MapConfig(fullConfig), REWRITER_NAME);
-  }
-
-  @Test(expected = SamzaException.class)
-  public void testApplyRewriterClassDoesNotExist() {
-    Map<String, String> fullConfig =
-        ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, String.format(JobConfig.CONFIG_REWRITER_CLASS, REWRITER_NAME),
-            "not_a_class");
-    Config expectedConfig = new MapConfig(ImmutableMap.of(CONFIG_KEY, CONFIG_VALUE, NEW_CONFIG_KEY, CONFIG_VALUE));
-    assertEquals(expectedConfig, Util.applyRewriter(new MapConfig(fullConfig), REWRITER_NAME));
-  }
-
   /**
    * No requirement for this test that this extends any other class. Just need some placeholder class.
    */
@@ -294,48 +179,5 @@ public class TestUtil {
    * No requirement for this test that this extends any other class. Just need some placeholder class.
    */
   public static class MyTaskClass {
-  }
-
-  /**
-   * Adds a new config entry for the key {@link #NEW_CONFIG_KEY} which has the same value as {@link #CONFIG_KEY}.
-   */
-  public static class NewPropertyRewriter implements ConfigRewriter {
-    @Override
-    public Config rewrite(String name, Config config) {
-      ImmutableMap.Builder<String, String> newConfigMapBuilder = new ImmutableMap.Builder<>();
-      newConfigMapBuilder.putAll(config);
-      newConfigMapBuilder.put(NEW_CONFIG_KEY, config.get(CONFIG_KEY));
-      return new MapConfig(newConfigMapBuilder.build());
-    }
-  }
-
-  /**
-   * If an entry at {@link #NEW_CONFIG_KEY} exists, overwrites it to be the value concatenated with itself. Otherwise,
-   * updates the entry at {@link #CONFIG_KEY} to be the value concatenated to itself.
-   */
-  public static class UpdatePropertyRewriter implements ConfigRewriter {
-    @Override
-    public Config rewrite(String name, Config config) {
-      Map<String, String> newConfigMap = new HashMap<>(config);
-      if (config.containsKey(NEW_CONFIG_KEY)) {
-        // for testing overwriting of new configs
-        newConfigMap.put(NEW_CONFIG_KEY, config.get(NEW_CONFIG_KEY) + config.get(NEW_CONFIG_KEY));
-      } else {
-        newConfigMap.put(CONFIG_KEY, config.get(CONFIG_KEY) + config.get(CONFIG_KEY));
-      }
-      return new MapConfig(newConfigMap);
-    }
-  }
-
-  /**
-   * Removes config entry for the key {@link #CONFIG_KEY} and {@link #NEW_CONFIG_KEY}.
-   */
-  public static class DeletePropertyRewriter implements ConfigRewriter {
-    @Override
-    public Config rewrite(String name, Config config) {
-      Map<String, String> newConfigMap = new HashMap<>(config);
-      newConfigMap.remove(CONFIG_KEY);
-      return new MapConfig(newConfigMap);
-    }
   }
 }
